third_party/WebKit/LayoutTests/fast/dom/shadow/get-destination-insertion-points-skips-user-agent-shadow-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that getDestinationInsertionPoints() should skip insertion points in user-agent shadow roots.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS getNodeInComposedTree('host/detail-child').getDestinationInsertionPoints().length is 0
PASS successfullyParsed is true

TEST COMPLETE
